+++ Assets/scripts/modding/scripts/home.js
@@ -1,0 +1,4 @@
+import { createImage, createParagraph, replaceText } from "../api.js";
+
+replaceText("title", 'welcome to the dragging modded site!');
+createParagraph('isjfghydfnkchcxhnkfvghjcfdhyh');