+++ script-packets/modding/template/404.js
@@ -1,5 +1,0 @@
-import { replaceImg, replaceText } from '../api.js';
-
-// Text
-replaceText('name', '');
-replaceText('nonexist', '');